Refactor memmap dataset tests so that testing of neighbor functions lives separately from core implementation testing
